compiler/semantic: enable type checking for all input files
For file types with an sio.Reader that has a Type method, continue to
use that to obtain a super.Type.  For other file types, if the input can
seek, read all values from the file and fuse their types to obtain a
single type.  (Fifos and pipes are excluded for now because the runtime
must also be able to read the file.)

To disable type checking for all input files, specify the -dynamic flag
to the super and super compile commands.