switch to pragma-controlled Postgres style relational scoping
This commit changes the relational scoping rules to follow Postgres
except for GROUP BY expressions which check column aliases before
the input table (like Google SQL).  This is favored as the default
for SuperSQL because it doesn't change binding when a dynamic input
transitions to a typed input (and thus a field that was present but
unknown in the input changes the GROUP BY binding of a column alias).

This behavior can be overridden now with "pragma pg=true" which says
to follow Postgres semantics and favor the input table over any column
alias in GROUP BY expressions.

As a result, column aliases are no longer visible in WHERE clauses
or SELECT expressions in line with Postgres semantics (and ANSI
and Google SQL).

We added some light documentation to the pragma section of the book
in this PR, but more details and an explanation of the rationale for
this design will be forthcoming in a future PR from the book-sql branch.