new approach to relational scoping
This commit makes a big change to how relational scoping is implemented
better aligning the design with SQL semantics.  In this approach,
SQL operators always produce tables even if the input is not tabular.

Lateral column alias resolution is now performed statically and semantically
defined as substitution of input-scoped expressions corresponding
to the column alias.  To ensure that scope bindings do not depend on
input data, we chose to prioritize lateral column binding over input
table binding thus deviating from Postgres semantics in favor of
Google SQL semantics.

The select-* pattern no longer works with dynamic inputs and instead
matches all columns in the input schema.  Also, table.* pattern matching
now works.  We also fixed join with the USING clause to limit the USING
expressions to non-dotted identifiers and to deduplicate the USING columns
and place them in the leftmost column position.

Table and column scope bindings are now case insensitive so that
"select A" works on column "a" and "select t.a" works on table "T", etc.

Redundant agg functions are now detected and computed only once.  Also,
expressions to ORDER BY can correctly reach into the SELECT scope and
refer to aggregate functions, grouping expressions, input table columns,
and so forth.

We've updated the pipe join docs here to reflect the change to USING but
more complete doc updates for relational will be added in a subsequent PR.